requirements.txt
jedi==0.19.1
omgifol==0.5.1
packaging==24.1
parso==0.8.4
Pillow==9.5.0
Pygments==2.18.0
pypresence==4.3.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
typing_extensions==4.12.2
urwid==2.6.15
urwid_readline==0.14
wcwidth==0.2.13
winshell==0.6
